+++ CHANGELOG.md
@@ -3,7 +3,7 @@
 
 ## 0.4.15
 
-- Bump `@ecies/ciphers` to support 16 bytes nonce aes-256-gcm on deno
+- Bump dependencies
 - Revamp documentation
 
 ## 0.4.14